Patch from PR #25380
